Added `YoutubePanel`
