+++ src/components/editpanel.component.js
@@ -31,6 +31,7 @@ import {colours, styles} from "../pages/profileDesigns/colour.util";
 import {IoIosList, IoMdAdd, IoMdCloudUpload} from "react-icons/io";
 import GenericPanel from "./panels/generic.panel.component";
 import LinkListPanel from "./panels/linklist.panel.component";
+import YoutubePanel from "./panels/youtube.panel.component";
 
 const importAll = (r) => r.keys().map(r);
 const postFiles = importAll(require.context("../news/", true, /\.md$/))
@@ -59,16 +60,13 @@ export default class EditPanel extends React.Component
             // Spotify
             spotifyMessage: null,
             spotifyLink: "",
-            // YouTube
-            youtubeMessage: null,
-            youtubeLink: "",
+
         }
 
         this.handleLinkFieldChange = this.handleLinkFieldChange.bind(this)
         this.handleDisplayNameChange = this.handleDisplayNameChange.bind(this)
 
         this.handleSpotifyLinkChange = this.handleSpotifyLinkChange.bind(this)
-        this.handleYouTubeLinkChange = this.handleYouTubeLinkChange.bind(this)
     }
 
     async componentDidMount()
@@ -219,11 +217,6 @@ export default class EditPanel extends React.Component
         setTimeout(() => this.setState({spotifyMessage: null}), 5000)
     }
 
-    displayYouTubeMessage = (message) =>
-    {
-        this.setState({youtubeMessage: message})
-        setTimeout(() => this.setState({youtubeMessage: null}), 5000)
-    }
 
     drawMessage(message)
     {
@@ -340,12 +333,6 @@ export default class EditPanel extends React.Component
     {
         this.setState({spotifyLink: event.target.value})
     }
-
-    handleYouTubeLinkChange(event)
-    {
-        this.setState({youtubeLink: event.target.value})
-    }
-
     onProfilePictureChange = (event) =>
     {
         const allowedFiles = ['jpg', 'jpeg', 'png']
@@ -391,19 +378,6 @@ export default class EditPanel extends React.Component
         {
         })
     }
-
-    updateYouTubeLink = (link) =>
-    {
-        const regex = /^((?:https?:)?\/\/)?((?:www|m)\.)?((?:youtube(-nocookie)?\.com|youtu.be))(\/(?:[\w\-]+\?v=|embed\/|live\/|v\/)?)([\w\-]+)(\S+)?$/
-        let match = link.match(regex)
-        if (!match)
-            return this.displayYouTubeMessage({type: 'error', message: 'The provided YouTube link is invalid.'})
-
-        this.props.updateLocallyWithoutCancelling(match[6]).then(result =>
-        {
-        })
-    }
-
     setLinkListVertical = (vertical) =>
     {
         let content = this.props.selectedComponent.content
@@ -584,29 +558,12 @@ export default class EditPanel extends React.Component
                                       updateLocallyWithoutCancelling={this.props.updateLocallyWithoutCancelling}
                 />
             case 'youtube':
-                return <>
-                    <h3 className="m p-no-margin-top p-no-margin-bottom">Edit YouTube video</h3>
-                    {this.drawMessage(this.state.youtubeMessage)}
-                    <h2 className="s p-no-margin-bottom p-no-margin-top title">Import a YouTube video link:</h2>
-                    <input className="input" type="text" placeholder="https://www.youtube.com/watch?v=DgKpLoz29jo"
-                           value={this.state.youtubeLink} onChange={this.handleYouTubeLinkChange}/>
-                    <div>
-                        <button className="load-button"
-                                onClick={() => this.updateYouTubeLink(this.state.youtubeLink)}>Load video
-                        </button>
-                    </div>
-                    <iframe style={{borderRadius: "12px"}}
-                            src={"https://www.youtube-nocookie.com/embed/" + component.content}
-                            width={"100%"} height={400} frameBorder={"0"} allowFullScreen={true}
-                            allow={"autoplay; clipboard-write; encrypted-media; fullscreen; picture-in-picture"}
-                            loading={"lazy"}></iframe>
-                    <div className="margin-button">
-                        <button className="delete-component"
-                                onClick={() => this.props.deleteSelectedComponent()}>Delete component
-                        </button>
-                        <button className="done-button" onClick={() => this.cancel()}>Done</button>
-                    </div>
-                </>
+                return <YoutubePanel component={this.props.selectedComponent}
+                                      drawMessage={this.drawMessage}
+                                      deleteSelectedComponent={this.props.deleteSelectedComponent}
+                                      cancel={this.cancel} saveLocally={this.saveLocally}
+                                      updateLocallyWithoutCancelling={this.props.updateLocallyWithoutCancelling}
+                />
             case "spotify":
                 return <>
                     <h3 className="m p-no-margin-top p-no-margin-bottom">Edit Spotify playlist</h3>
